backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
numpy==2.1.1
scipy==1.14.1
pillow==10.4.0
torch==2.4.1
google-generativeai==0.8.3
python-dotenv==1.0.0
